requirements.txt
# Synnex Traffic Management System — Dependencies
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
#
# Install: pip install -r requirements.txt
#
# NOTE for Raspberry Pi:
#   picamera2 and opencv are installed via apt (system packages),
#   NOT pip, because building from source on Pi Zero 2W is very slow.
#   Use: sudo apt install python3-picamera2 python3-opencv
#   Then create venv with --system-site-packages:
#     python3 -m venv venv --system-site-packages

flask>=2.3.0
RPi.GPIO>=0.7.1
numpy>=1.24.0
opencv-python>=4.7.0; platform_machine != "armv7l" and platform_machine != "aarch64"
